Remove unused code from previous data extractors
