+++ milli/src/update/index_documents/mod.rs
@@ -1,24 +1,23 @@
 mod enrich;
-mod extract;
 mod extract2;
 mod helpers;
 mod transform;
-mod typed_chunk;
 
+use grenad::MergerBuilder;
 use heed::types::{ByteSlice, Str};
-use heed::BytesDecode;
 use heed::Database;
+use heed::{BytesDecode, RwTxn};
 use log::debug;
 use roaring::RoaringBitmap;
 use serde::{Deserialize, Serialize};
 use slice_group_by::GroupBy;
+use std::borrow::Cow;
 use std::collections::HashSet;
 use std::convert::TryInto;
-use std::io::{Cursor, Read, Seek};
+use std::io::{self, Cursor, Read, Seek};
 use std::iter::FromIterator;
 use std::num::{NonZeroU32, NonZeroUsize};
 use std::result::Result as StdResult;
-use typed_chunk::TypedChunk;
 
 use self::enrich::enrich_documents_batch;
 pub use self::enrich::{
@@ -32,11 +31,10 @@ pub use self::helpers::{
     sorter_into_lmdb_database, valid_lmdb_key, write_into_lmdb_database, writer_into_reader,
     ClonableMmap, MergeFn,
 };
-use self::helpers::{roaring_bitmap_from_u32s_array, GrenadParameters};
-pub use self::transform::{Transform, TransformOutput};
-use self::typed_chunk::{
-    append_entries_into_database, merge_word_docids_reader_into_fst, write_entries_into_database,
+use self::helpers::{
+    merge_ignore_values, roaring_bitmap_from_u32s_array, serialize_roaring_bitmap, GrenadParameters,
 };
+pub use self::transform::{Transform, TransformOutput};
 use crate::documents::{obkv_to_object, DocumentsBatchReader};
 use crate::error::UserError;
 use crate::heed_codec::facet::{decode_prefix_string, encode_prefix_string};
@@ -46,8 +44,8 @@ use crate::update::{
     WordPrefixPairProximityDocids, WordPrefixPositionDocids, WordsPrefixesFst,
 };
 use crate::{
-    lat_lng_to_xyz, BoRoaringBitmapCodec, GeoPoint, Index, Result, RoaringBitmapCodec,
-    SerializationError,
+    lat_lng_to_xyz, BoRoaringBitmapCodec, CboRoaringBitmapCodec, GeoPoint, Index, Result,
+    RoaringBitmapCodec, SerializationError,
 };
 
 static MERGED_DATABASE_COUNT: usize = 7;
@@ -364,7 +362,7 @@ where
                 self.wtxn,
                 index_is_empty,
                 |value, _buffer| Ok(value),
-                crate::update::index_documents::typed_chunk::merge_cbo_roaring_bitmaps,
+                merge_cbo_roaring_bitmaps_buffer,
             )?;
         }
         // NewDocumentsIds (nothing to do)
@@ -377,7 +375,7 @@ where
                 self.wtxn,
                 index_is_empty,
                 |value, _buffer| Ok(value),
-                crate::update::index_documents::typed_chunk::merge_roaring_bitmaps,
+                merge_roaring_bitmaps_buffer,
             )?;
 
             let exact_word_docids = unsafe { as_cloneable_grenad(&exact_word_docids) }?;
@@ -387,7 +385,7 @@ where
                 self.wtxn,
                 index_is_empty,
                 |value, _buffer| Ok(value),
-                crate::update::index_documents::typed_chunk::merge_roaring_bitmaps,
+                merge_roaring_bitmaps_buffer,
             )?;
 
             // create fst from word docids
@@ -410,7 +408,7 @@ where
                 self.wtxn,
                 index_is_empty,
                 |value, _buffer| Ok(value),
-                crate::update::index_documents::typed_chunk::merge_cbo_roaring_bitmaps,
+                merge_cbo_roaring_bitmaps_buffer,
             )?;
         }
         // FieldIdFacetNumberDocids
@@ -421,7 +419,7 @@ where
                 self.wtxn,
                 index_is_empty,
                 |value, _buffer| Ok(value),
-                crate::update::index_documents::typed_chunk::merge_cbo_roaring_bitmaps,
+                merge_cbo_roaring_bitmaps_buffer,
             )?;
         }
         // FieldIdFacetExistsDocids
@@ -432,7 +430,7 @@ where
                 self.wtxn,
                 index_is_empty,
                 |value, _buffer| Ok(value),
-                crate::update::index_documents::typed_chunk::merge_cbo_roaring_bitmaps,
+                merge_cbo_roaring_bitmaps_buffer,
             )?;
         }
         // WordPairProximityDocids
@@ -445,7 +443,7 @@ where
                 self.wtxn,
                 index_is_empty,
                 |value, _buffer| Ok(value),
-                crate::update::index_documents::typed_chunk::merge_cbo_roaring_bitmaps,
+                merge_cbo_roaring_bitmaps_buffer,
             )?;
         }
         // FieldIdDocidFacetNumbers
@@ -714,6 +712,128 @@ fn execute_word_prefix_docids(
         &common_prefix_fst_words,
         &del_prefix_fst_words,
     )?;
+    Ok(())
+}
+
+pub fn merge_word_docids_reader_into_fst(
+    word_docids_iter: grenad::Reader<io::Cursor<ClonableMmap>>,
+    exact_word_docids_iter: grenad::Reader<io::Cursor<ClonableMmap>>,
+) -> Result<fst::Set<Vec<u8>>> {
+    let mut merger_builder = MergerBuilder::new(merge_ignore_values as MergeFn);
+    merger_builder.push(word_docids_iter.into_cursor()?);
+    merger_builder.push(exact_word_docids_iter.into_cursor()?);
+    let mut iter = merger_builder.build().into_stream_merger_iter()?;
+    let mut builder = fst::SetBuilder::memory();
+
+    while let Some((k, _)) = iter.next()? {
+        builder.insert(k)?;
+    }
+
+    Ok(builder.into_set())
+}
+
+pub fn merge_roaring_bitmaps_buffer(
+    new_value: &[u8],
+    db_value: &[u8],
+    buffer: &mut Vec<u8>,
+) -> Result<()> {
+    let new_value = RoaringBitmap::deserialize_from(new_value)?;
+    let db_value = RoaringBitmap::deserialize_from(db_value)?;
+    let value = new_value | db_value;
+    Ok(serialize_roaring_bitmap(&value, buffer)?)
+}
+
+pub fn merge_cbo_roaring_bitmaps_buffer(
+    new_value: &[u8],
+    db_value: &[u8],
+    buffer: &mut Vec<u8>,
+) -> Result<()> {
+    Ok(CboRoaringBitmapCodec::merge_into(
+        &[Cow::Borrowed(db_value), Cow::Borrowed(new_value)],
+        buffer,
+    )?)
+}
+
+/// Write provided entries in database using serialize_value function.
+/// merge_values function is used if an entry already exist in the database.
+pub fn write_entries_into_database<R, K, V, FS, FM>(
+    data: grenad::Reader<R>,
+    database: &heed::Database<K, V>,
+    wtxn: &mut RwTxn,
+    index_is_empty: bool,
+    serialize_value: FS,
+    merge_values: FM,
+) -> Result<()>
+where
+    R: io::Read + io::Seek,
+    FS: for<'a> Fn(&'a [u8], &'a mut Vec<u8>) -> Result<&'a [u8]>,
+    FM: Fn(&[u8], &[u8], &mut Vec<u8>) -> Result<()>,
+{
+    let mut buffer = Vec::new();
+    let database = database.remap_types::<ByteSlice, ByteSlice>();
+
+    let mut cursor = data.into_cursor()?;
+    while let Some((key, value)) = cursor.move_on_next()? {
+        if valid_lmdb_key(key) {
+            buffer.clear();
+            let value = if index_is_empty {
+                serialize_value(value, &mut buffer)?
+            } else {
+                match database.get(wtxn, key)? {
+                    Some(prev_value) => {
+                        merge_values(value, prev_value, &mut buffer)?;
+                        &buffer[..]
+                    }
+                    None => serialize_value(value, &mut buffer)?,
+                }
+            };
+            database.put(wtxn, key, value)?;
+        }
+    }
+
+    Ok(())
+}
+
+/// Write provided entries in database using serialize_value function.
+/// merge_values function is used if an entry already exist in the database.
+/// All provided entries must be ordered.
+/// If the index is not empty, write_entries_into_database is called instead.
+pub fn append_entries_into_database<R, K, V, FS, FM>(
+    data: grenad::Reader<R>,
+    database: &heed::Database<K, V>,
+    wtxn: &mut RwTxn,
+    index_is_empty: bool,
+    serialize_value: FS,
+    merge_values: FM,
+) -> Result<()>
+where
+    R: io::Read + io::Seek,
+    FS: for<'a> Fn(&'a [u8], &'a mut Vec<u8>) -> Result<&'a [u8]>,
+    FM: Fn(&[u8], &[u8], &mut Vec<u8>) -> Result<()>,
+{
+    if !index_is_empty {
+        return write_entries_into_database(
+            data,
+            database,
+            wtxn,
+            false,
+            serialize_value,
+            merge_values,
+        );
+    }
+
+    let mut buffer = Vec::new();
+    let mut database = database.iter_mut(wtxn)?.remap_types::<ByteSlice, ByteSlice>();
+
+    let mut cursor = data.into_cursor()?;
+    while let Some((key, value)) = cursor.move_on_next()? {
+        if valid_lmdb_key(key) {
+            buffer.clear();
+            let value = serialize_value(value, &mut buffer)?;
+            unsafe { database.append(key, value)? };
+        }
+    }
+
     Ok(())
 }
